verify_bot_app/requirements.txt
python-telegram-bot
pyTelegramBotAPI
uvicorn